refactor(api): extract booking handlers from mod.rs into bookings.rs
Move 1502 lines of booking handler code from api/mod.rs into the
dedicated bookings.rs module (which previously only contained tests).

Extracted handlers:
- list_bookings, create_booking, get_booking, cancel_booking
- get_booking_invoice (HTML invoice generation)
- quick_book, booking_checkin, update_booking
- QuickBookRequest, PatchBookingRequest structs

mod.rs: 5957 -> 4460 lines (25% reduction)
bookings.rs: 265 -> 1791 lines (handlers + existing tests)

All 764 tests pass. Router, OpenAPI references updated.
